requirements.txt
streamlit
pytesseract
pillow
rapidfuzz
python-dotenv
groq
easyocr
torch